Implement AloTra storefront with JPA and Sitemesh layout

+++ src/main/java/dao/CategoryDao.java
@@ -2,12 +2,12 @@ package dao;
 import java.util.List;
 import model.Category;
 
-public interface CategoryDao {
-  void insert(Category c);
-  void edit(Category c);
-  void delete(int id);
-  Category get(int id);
-  Category get(String name);
-  List<Category> getAll();
-  List<Category> search(String keyword);
-}
+  public interface CategoryDao {
+    void insert(Category c);
+    void edit(Category c);
+    void delete(Long id);
+    Category get(Long id);
+    Category get(String name);
+    List<Category> getAll();
+    List<Category> search(String keyword);
+  }
